+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### 12.57.0
+
+* `[log]` Added interface for compatible loggers
+* `[httputil]` Added more helpers for checking URL scheme
+* `[log]` Color for critical errors set from magenta to bold red
+* `[tmp]` Improved unique name generation
+* `[req]` Code refactoring
+* `[tmp]` Improved usage examples
+
 ### 12.56.0
 
 * `[fmtc]` Added named colors support